requirements.txt
torchvision==0.18.1+cu118
pillow==10.3.0
numpy==1.26.4
matplotlib==3.7.2